Move PackageSpec to binstar_client/utils/spec.py
